apicon/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-forms-gds==0.2.4
Django==4.2
django-crispy-forms==2.0
django-cron==0.6.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
pipreqs==0.4.12
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
tzlocal==2.1
urllib3==1.26.15
wfastcgi==3.0.0
yarg==0.1.9
